Remove redundent codes in benchmark tests
